+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ### 1.7.9
 * refactored controller code base
 * refactored account templates
+* allowing access to plugin classes
 * bugfixes
 
 ### 1.7.8